Path Tracing integrator: added new Russian Roulette parameter to speed up path tracing
As requested in http://www.yafaray.org/node/775, I've added a basic russian roulette to the path tracing as well as a new render boolean parameter "russian_roulette" to control it. From my tests it seems a lot faster (it renders in less than half the time of normal path tracing) but it looks noisier to me. Eventually more samples have to be used and the speed gain will not be that good, but still perhaps might get some speed improvement for the same image quality...