IMPORTANT: all integrators, SPPM and path roulette: Fixing non-randomness repetitive patterns As described in http://www.yafaray.org/node/792 I discovered lack of randomness in the bidirectional integrator.
I found out that it was caused by an insufficiently random "seeding" of the random_t pnrg object. For example during first AA pass, offset = 0 so it the pnrg object was seeded with value "123" for each tile causing the artifact detected in the bidir integrator.

However, and even when it's not obviously visible, I also believe it affected the SPPM integrator as well because of the same issue. Therefore I proceeded to add more randomness to the pnrg seeding for all integrators, including SPPM. I hope this also improves russian roulette randomness.

This change should not cause ill effects and should be beneficial in my opinion, but it's difficult to know for sure. We need to keep an eye on this to ensure no new issues happen now.