IMPORTANT: Path/Photon OneDirectLight - attempt to sample lights more uniformly
As reported in http://www.yafaray.org/node/803 there are artifacts in the path tracing and photon mapping algorithms when there is more than one light.

I found that the allegedly "random" sampling of lights in the mcintegrator.cc "estimateOneDirectLight" was not sampling all lights in the scene with the same probability. It was (incorrectly I think) sampling more often the first light, and less often the last light in the lights vector. As the result of each sample is multiplied by the number of lights, it caused the lighting to be non-uniform.

First I attempted to use a "randomized" approach to solve this, but it increased the noise (not too much, but it was noticeable). So I decided to try to improve the sampling using the current Halton sampling algorithm. I found that the input to the algorithm was not correlative enough, having many "jumps" that probably caused the wrong Halton outputs. Therefore I decided to change the input data and use an almost purely correlative numbering for each sample (having separate correlative numbers, one per thread). This seems to improve the noise and still give similar results no matter how many lights are used, and a more uniform light sampling.

This is a very significant change. I would expect scenes to be a little bit noisier now, but having a much more correct lighting when there is more than one light in the scene.

In any case we need to keep an eye on this and perhaps fine tune it more or even revert it back completely and look for another solution. For now I will leave it this way and see what happens...