Compatibility fix for Python 3.8
In Python 3.8, the reserved "tp_print" slot was changed from
a function pointer to a number, `Py_ssize_t tp_vectorcall_offset`.
In C, there is no "nullptr"; either a 0 or NULL casts automatically
to both pointers and numbers.

Use 0 instead of "nullptr" to be source-compatible both with
Python 3.8 and previous versions.