MAJOR: Vector: Fixed *difficult* bug randomly affecting texture mapping in some compilers like CLang * When building with some compilers such as CLang in Linux, in Debug mode it worked normally but in Release or RelWithDebugInfo modes it caused bad corruption in the texture mapping. * When profiled, I found the issue was caused by uninitialized values. Chasing the problem I found a possible cause at the TextureMapperNode::cubeMap function. * This was difficult because there was nothing clearly causing it. When adding some std::cout code to the cubeMap function, it started to work correctly which pointed to a problem caused/triggered by the optimization of the function. * Finally I realized, after a long painful investigation, that the problem was likely caused by these functions at Vec3 class: 	float operator[](int i) const { return (&x_)[i]; } //Lynx 	float &operator[](int i) { return (&x_)[i]; } //Lynx * I realized that those very old "hacks" were ok in most cases but apparently in some cases like CLang in Release mode, it was packing the Vec3 internal structure made of float x_, y_, z_ in a different way, so directly accessing the float members by address was causing undefined behavior and the uninitialized values
I spotted similar "hackish" code in the Vec3Double class.

Therefore I've had to change the entire code, either to avoid direct numeric access to the members (which proved to be unfeasible as some functions needed this), or else to change the internal Vec3 and Vec3Double classes to a proper contiguous array using std::array

So I had to pursue the second option. This has the downside that directly accessing x_, y_, z_ is no longer possible so I have to provide new functions to access those data members instead of doing it directly, with really big changes to the entire codebase.