MAJOR: unified scene items parameters and metadata system This is a huge commit containing major changes in the entire YafaRay scene items creation system, and a number of other changes and old/unused code cleanup. It is also a pre-requirement (among others) that had to be done in order to allow the preparation of the final libYafaRay v4.0 API (work in progress). Only when the final libYafaRay v4.0 API is finalized in the future we will be able to change from pre-alpha status to alpha status.
YafaRay scene items creation (such as materials, cameras, objects, lights, etc) has always been done using a flexible "runtime parameters" system. That is useful to keep a simple public API without the need to change scene items creation function signatures or to use structures in the API. However the way it was implemented had a number of significant downsides:
* There was no way to obtain lists of available parameters for the available scene items, including the description of each parameter's function and usage. This lack of "scene items/parameters metadata" has always been a problem specially with the lack of scene items parameters documentation.
* YafaRay did not check if a certain parameter that was used during scene item creation was actually used/recognized or not.
* There was no check to ensure that the type of the parameter (Int, Float, String, etc) actually used during scene item creation was actually correct for the parameter expected by the scene item
* The parameters used during scene items creation were not stored with the scene items. This made it impossible to save an existing scene that was created dynamically (for example in Python)
* Scene items were created using factories and constructors, but there were several problems with the way they were implemented:
- Variables were initialized in many different places (class variables, factory variables, constructors), with a lot of code repetition
- It was difficult to understand exactly how variables/parameters were used and their function in the scene item being created, due to complex initialization and preprocessing
- Construction encapsulation was often broken, with factories creating the class instance of the scene item and afterwards accessing directly some of the class members to complete the scene item initialization. This model had the disadvantages of a rigid construction and did not have the advantages of construction encapsulation.

The changes are indeed deep and complex. Usage of macros is extensive in the new code for the parameters definition, initialization, saving and checking operations. I always try to avoid macros and replace them by more modern C++ alternatives, but due to the lack of C++17 reflection functionality, after a long time researching possible alternatives I came to the conclussion that without macros, the code for the new parameters/metadata system would have been too convoluted and difficult to write, read and maintain.

I'm afraid that there is no perfect system and we always need some kind of trade-off. In this case the objectives/criteria to decide how to implement the changes have been as follows:
* Create a unified way to declare the list of available parameters for each class / scene item, including the type, name, description, and default value
* Create a unified way to declare and use enumerations (strings that contain values to be chosen from a list), without having to always create ad-hoc code to convert between the enumerated strings used in the parameter to the actual enum values used in the code
* Create a way to obtain lists of available parameters for a certain scene item class, including their type, name, description and default value (or list of available enumerated options in enums)
* Create a way to export the actual parameters used in an object to a parameter map, to allow future scene saving operations. This export operation allows to export all parameters or only the parameters that have been changed respect to their default values.
* Create a unified way to load and save the parameters and enumeration parameters in each instance of a scene item, including proper "cascading" of parameters declaration/loading/saving for scene items that are derived from other "parent" scene item classes.
* Streamline the instance construction with almost direct creation from the list of parameters, showing clearly how the parameters are used to construct the instance and the relationships between object construction and the parameters
* Minimize the size increase and speed slowdowns in the classes due to the new parameters metadata system, by:
- Using static class members for big parameters metadata such as name, description, default value, that do not need to be wasted in each instance
- Using unsigned char for Enum values (leaving open the possibility to use bigger sizes in the future like unsigned short if more than 8 bits are needed)
- Using classic C++ member variables instead of keeping directly parameter maps, for avoiding slowing down finding and using the values. This has a cost in code complexity, where for loading or saving parameters from/to parameter maps, a conversion must take place.
- Using macros to reduce the code complexity to declare parameters and load/save them into class members.

The trade-off in this case is that the constructors are no longer simple/well defined/strict. The constructors now take the parameter map as an argument, so they are not formally correctly encapsulated anymore. However, before these changes, although they were partially encapsulated, factories often accessed the class members directly after construction which broke encapsulation anyway and made objects initialization rather confusing.

In addition to those changes, there are a number of class relocations. I've tried to allocate each class in its own file (header and sometimes implementation file) to further clarify the code and allow easier maintenance.

I've also made a number of other changes like removing unnecessary code and other minor changes at the same time as the big changes.
- Most importantly, I've changed some of the parameter names (which will need to also be changed in the Blender exporter, for example)
- Changes to ImageBuffers code, simplifying classes using templates
- Added C API functions to get image dimensions, fixed some C test example parameters, etc.
- Unifiying flags handling and changes to Visibility
- Fixed DarkSky: removing parameters that had been previously disabled/hardcoded by darktide

All this should still be considered a "work in progress", still lots of work to do to fine tuning the system, possibly further parameter names might be changed and many of the parameter descriptions are still empty and should be written for properly documented the classes.